requirements.txt
pycryptodome
flask
flask-cors
opencv-python
numpy
binascii
nearpy
redis
pandas
seaborn